requirements.txt
Django==4.0
Pillow
numpy
pandas
tensorflow-hub
tensorflow==2.13.0
protobuf==3.20.3
openai==0.28
python-dotenv
scikit-learn
webcolors
